Conflicts:
	kernel/src/main/java/org/neo4j/kernel/impl/nioneo/xa/WriteTransaction.java
	kernel/src/test/java/org/neo4j/kernel/impl/nioneo/store/TestNeoStore.java

+++ kernel/src/main/java/org/neo4j/kernel/impl/nioneo/xa/WriteTransaction.java
@@ -755,7 +755,7 @@ public class WriteTransaction extends XaTransaction implements NeoStoreTransacti
             PrimitiveRecord primitive )
     {
         ArrayMap<Integer, PropertyData> result = new ArrayMap<Integer, PropertyData>(
-                9, false, true );
+                (byte)9, false, true );
         long nextProp = primitive.getFirstProp();
         while ( nextProp != Record.NO_NEXT_PROPERTY.intValue() )
         {
